feat: add modal for terms of use
