--- src/components/icons/logo.tsx
@@ -1,52 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-const LogoIcon = createIcon({
-  displayName: 'LogoIcon',
-  viewBox: '0 0 200 75',
-  path: (
-    <svg viewBox="0 0 200 78" fill="blue" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M162.324 59.3184C162.297 59.2665 162.283 59.2003 162.283 59.1216C162.283 58.991 162.372 58.898 162.551 58.8461L163.086 58.669L163.272 58.6297C163.395 58.6297 163.499 58.7155 163.581 58.8855C163.814 59.4883 164.191 59.9248 164.714 60.1932C165.234 60.4615 165.907 60.5957 166.73 60.5957C167.703 60.5957 168.431 60.4418 168.911 60.1341C169.39 59.8265 169.632 59.3828 169.632 58.8068C169.632 58.3613 169.512 58.0161 169.272 57.7728C169.032 57.5295 168.708 57.356 168.304 57.2504C167.9 57.1449 167.347 57.0608 166.648 56.9946C165.755 56.9159 165.029 56.8014 164.467 56.6511C163.903 56.5009 163.435 56.2182 163.057 55.805C162.68 55.3918 162.492 54.7978 162.492 54.0232C162.492 53.1055 162.839 52.3828 163.531 51.8587C164.224 51.3345 165.229 51.0715 166.546 51.0715C168.685 51.0715 170.086 51.7871 170.744 53.2164C170.785 53.2952 170.807 53.3667 170.807 53.4329C170.807 53.5635 170.725 53.6619 170.56 53.7281L170.045 53.9445C169.948 53.9839 169.873 54.0036 169.818 54.0036C169.762 54.0036 169.71 53.9803 169.664 53.9356C169.615 53.8891 169.571 53.8336 169.53 53.7674C169.295 53.2683 168.954 52.8926 168.501 52.635C168.048 52.3792 167.403 52.2522 166.567 52.2522C165.674 52.2522 165.004 52.4007 164.551 52.6941C164.098 52.9892 163.871 53.4186 163.871 53.9839C163.871 54.4687 163.998 54.8372 164.252 55.0858C164.506 55.3345 164.835 55.5062 165.24 55.5975C165.644 55.6887 166.224 55.7674 166.979 55.8336C167.843 55.9123 168.556 56.0375 169.118 56.2075C169.68 56.3774 170.131 56.6637 170.467 57.0644C170.803 57.4651 170.971 58.025 170.971 58.7477C170.971 59.6923 170.617 60.4311 169.912 60.9624C169.206 61.4937 168.152 61.7603 166.753 61.7603C164.462 61.7603 162.988 60.9463 162.329 59.3202L162.324 59.3184Z"
-        fill="#FFC010"
-      />
-      <path
-        d="M174.775 61.4131C174.129 61.1842 173.68 60.8496 173.426 60.4096C173.172 59.9695 173.045 59.4024 173.045 58.7083C173.045 57.948 173.195 57.3469 173.497 56.9069C173.8 56.4668 174.279 56.152 174.937 55.9623C175.596 55.7727 176.474 55.6779 177.572 55.6779C178.546 55.6779 179.445 55.7441 180.268 55.8747V54.9498C180.268 54.0965 180.066 53.4346 179.662 52.9623C179.258 52.4901 178.533 52.2539 177.49 52.2539C176.132 52.2539 175.165 52.7262 174.589 53.6707C174.547 53.7369 174.503 53.7924 174.454 53.8371C174.406 53.8836 174.354 53.9069 174.3 53.9069C174.247 53.9069 174.177 53.8872 174.095 53.8478L173.601 53.6511C173.422 53.5992 173.333 53.5062 173.333 53.3756C173.333 53.3094 173.354 53.2307 173.395 53.1394C174.163 51.762 175.542 51.0732 177.531 51.0732C179.054 51.0732 180.122 51.4239 180.732 52.1269C181.342 52.8282 181.648 53.7763 181.648 54.9695V60.1251C181.648 60.4668 181.596 60.7083 181.494 60.8532C181.39 60.9981 181.195 61.1287 180.907 61.2468C180.084 61.5884 178.945 61.7584 177.49 61.7584C176.324 61.7584 175.419 61.6439 174.773 61.4149L174.775 61.4131ZM175.063 60.1734C175.474 60.4954 176.249 60.6564 177.388 60.6564C178.528 60.6564 179.315 60.5509 179.878 60.3416C180.043 60.2754 180.148 60.211 180.197 60.1448C180.245 60.0786 180.268 59.9552 180.268 59.7709V56.9176C179.034 56.7995 178.088 56.7405 177.428 56.7405C176.329 56.7405 175.558 56.8908 175.113 57.1931C174.667 57.4954 174.444 57.9927 174.444 58.6886C174.444 59.3845 174.649 59.8532 175.061 60.1734H175.063Z"
-        fill="#FFC010"
-      />
-      <path
-        d="M185.723 61.4044C185.667 61.3525 185.64 61.2595 185.64 61.1289V52.5886H183.809C183.671 52.5886 183.574 52.5618 183.521 52.5099C183.465 52.4581 183.438 52.365 183.438 52.2344V51.7228C183.438 51.5779 183.465 51.4831 183.521 51.4366C183.576 51.3901 183.671 51.3686 183.809 51.3686H185.64V50.8767C185.64 49.4724 185.962 48.4688 186.608 47.8659C187.252 47.2631 188.317 46.9482 189.798 46.9214H189.962C190.113 46.9214 190.213 46.9446 190.261 46.9894C190.309 47.0359 190.333 47.1307 190.333 47.2756V47.7872C190.333 47.9321 190.306 48.0305 190.25 48.0824C190.195 48.1343 190.1 48.1611 189.962 48.1611C189.193 48.1611 188.601 48.2398 188.182 48.3972C187.764 48.5547 187.465 48.8266 187.288 49.2148C187.109 49.6012 187.02 50.1557 187.02 50.8767V51.3686H190.004C190.141 51.3686 190.236 51.3954 190.292 51.4473C190.347 51.4992 190.374 51.5922 190.374 51.7228V52.2344C190.374 52.365 190.347 52.4581 190.292 52.5099C190.236 52.5618 190.141 52.5886 190.004 52.5886H187.02V61.1289C187.02 61.2595 186.996 61.3525 186.948 61.4044C186.9 61.4563 186.8 61.4831 186.649 61.4831H186.011C185.873 61.4831 185.776 61.4563 185.723 61.4044Z"
-        fill="#FFC010"
-      />
-      <path
-        d="M192.886 60.6958C192.145 59.9874 191.775 59.016 191.775 57.7834V55.2253C191.775 53.914 192.163 52.8943 192.937 52.1663C193.712 51.4382 194.841 51.0732 196.322 51.0732C197.803 51.0732 198.954 51.4435 199.728 52.1859C200.503 52.9266 200.891 53.914 200.891 55.1466V56.2289C200.891 56.5956 200.685 56.7799 200.274 56.7799H193.152V57.9212C193.152 58.7745 193.417 59.4364 193.945 59.9087C194.472 60.3809 195.238 60.6171 196.24 60.6171C197.762 60.6171 198.839 60.0732 199.471 58.9838C199.594 58.8389 199.703 58.7673 199.8 58.7673C199.882 58.7673 199.971 58.7942 200.068 58.8461L200.501 59.0625C200.666 59.1412 200.748 59.2396 200.748 59.3577C200.748 59.4364 200.714 59.5276 200.644 59.6332C199.821 61.05 198.345 61.7584 196.22 61.7584C194.739 61.7584 193.626 61.4042 192.886 60.6958ZM199.512 55.7566V55.0679C199.512 54.1752 199.24 53.4775 198.7 52.9731C198.157 52.4686 197.365 52.2164 196.322 52.2164C195.279 52.2164 194.508 52.4686 193.966 52.9731C193.424 53.4775 193.154 54.177 193.154 55.0679V55.7566H199.513H199.512Z"
-        fill="#FFC010"
-      />
-      <path
-        d="M91.7407 61.1181C90.3167 60.619 89.3221 59.8837 88.7568 58.9124C88.1915 57.941 87.9106 56.7013 87.9106 55.1896C87.9106 53.517 88.2487 52.2022 88.9267 51.2451C89.6047 50.288 90.6477 49.6065 92.0573 49.2022C93.467 48.7979 95.3722 48.5957 97.7693 48.5957C98.5028 48.5957 99.5046 48.6369 100.773 48.7174V47.746C100.773 46.7209 100.519 45.9589 100.011 45.4598C99.5028 44.9607 98.6709 44.712 97.5135 44.712C95.9607 44.712 94.7908 45.2111 94.0019 46.2093C93.7765 46.5331 93.5779 46.7621 93.4098 46.898C93.2398 47.034 93.0287 47.1002 92.7747 47.1002C92.5493 47.1002 92.238 47.0197 91.8445 46.8569L89.5171 45.966C88.755 45.6691 88.374 45.2648 88.374 44.7514C88.374 44.4812 88.4724 44.1717 88.6709 43.8211C89.433 42.4723 90.5672 41.4347 92.077 40.7048C93.5851 39.9768 95.4688 39.6118 97.7264 39.6118C101.111 39.6118 103.53 40.3596 104.984 41.8569C106.437 43.3542 107.163 45.4508 107.163 48.1485V58.4258C107.163 59.127 107.057 59.6333 106.846 59.9428C106.635 60.254 106.218 60.5295 105.598 60.7728C103.905 61.5009 101.24 61.8658 97.5994 61.8658C95.1164 61.8658 93.1629 61.6154 91.7389 61.1181H91.7407ZM94.8928 56.6476C95.2881 56.9982 96.0359 57.1735 97.136 57.1735C98.5189 57.1735 99.6334 57.1056 100.48 56.9714C100.676 56.9177 100.776 56.8229 100.776 56.6887V53.2505C99.1683 53.17 97.984 53.1288 97.2219 53.1288C96.0931 53.1288 95.3239 53.3041 94.916 53.6548C94.5064 54.0054 94.3024 54.5045 94.3024 55.1521C94.3024 55.7997 94.4992 56.2988 94.8945 56.6494L94.8928 56.6476Z"
-        fill="#F5F5F5"
-      />
-      <path
-        d="M135.857 59.6207C134.136 58.1234 133.275 56.0393 133.275 53.3702V48.1502C133.275 45.4525 134.136 43.3559 135.857 41.8586C137.578 40.3613 140.073 39.6135 143.347 39.6135C146.621 39.6135 149.159 40.3613 150.88 41.8586C152.601 43.3559 153.461 45.4525 153.461 48.1502V53.3702C153.461 56.0679 152.601 58.1591 150.88 59.6421C149.159 61.1252 146.647 61.8676 143.347 61.8676C140.046 61.8676 137.578 61.1198 135.857 59.6225V59.6207ZM140.596 55.6547C141.216 56.1806 142.132 56.4436 143.347 56.4436C144.562 56.4436 145.519 56.1806 146.139 55.6547C146.76 55.1287 147.07 54.381 147.07 53.4096V48.0697C147.07 46.0464 145.828 45.0357 143.345 45.0357C140.862 45.0357 139.664 46.0464 139.664 48.0697V53.4096C139.664 54.381 139.973 55.1287 140.596 55.6547Z"
-        fill="#F5F5F5"
-      />
-      <path
-        d="M131.032 59.44L127.351 52.3184C126.617 50.9427 125.827 49.9713 124.981 49.406L130.693 42.0017C130.973 41.6243 131.116 41.2879 131.116 40.991C131.116 40.4794 130.608 40.2218 129.592 40.2218H126.757C125.854 40.2218 125.218 40.2897 124.852 40.4239C124.485 40.5599 124.118 40.8962 123.752 41.4346L119.013 47.8676H117.828V31.3184L111.438 35.0071V59.7209C111.438 60.3148 111.551 60.7191 111.777 60.9355C112.002 61.152 112.426 61.2593 113.047 61.2593H116.22C116.841 61.2593 117.263 61.152 117.49 60.9355C117.716 60.7191 117.828 60.3148 117.828 59.7209V53.288H119.31C120.015 53.288 120.58 53.4168 121.002 53.6726C121.426 53.9284 121.764 54.3398 122.018 54.9069L124.261 59.8837C124.542 60.4776 124.854 60.855 125.192 61.016C125.53 61.1788 126.15 61.2593 127.054 61.2593H129.973C130.481 61.2593 130.848 61.2057 131.074 61.0983C131.299 60.991 131.412 60.8014 131.412 60.5312C131.412 60.2343 131.285 59.8711 131.031 59.4382L131.032 59.44Z"
-        fill="#F5F5F5"
-      />
-      <path
-        d="M82.6655 41.7781C81.1414 40.3362 78.8569 39.6135 75.8104 39.6135C74.4848 39.6135 73.1575 39.7495 71.8319 40.0178V31.3184L65.4419 35.0071V58.2647C65.4419 58.9123 65.5331 59.3702 65.7174 59.6404C65.8999 59.9105 66.331 60.2074 67.0072 60.5312C67.882 60.9087 69.0233 61.2253 70.4347 61.4829C71.8444 61.7388 73.3686 61.8676 75.0054 61.8676C78.3059 61.8676 80.7889 61.1198 82.4526 59.6225C84.1163 58.1252 84.9499 55.9337 84.9499 53.0482V47.991C84.9499 45.2933 84.1879 43.2235 82.6655 41.7799V41.7781ZM78.56 53.0464C78.56 54.1788 78.2433 55.0232 77.6083 55.576C76.9732 56.1287 75.9929 56.406 74.6673 56.406C73.5671 56.406 72.6637 56.2987 71.9589 56.0822C71.8748 56.0554 71.8319 55.9874 71.8319 55.8801V45.5223C72.8748 45.2253 73.9195 45.0769 74.9625 45.0769C76.1181 45.0769 77.0072 45.3273 77.6279 45.8246C78.2487 46.3237 78.5582 47.059 78.5582 48.0303V53.0482L78.56 53.0464Z"
-        fill="#F5F5F5"
-      />
-      <path
-        d="M0.926514 38.8857L45.8371 61.2482L23.3863 25.9226L0.926514 38.8857Z"
-        fill="#FFC010"
-      />
-      <path
-        d="M40.0108 31.966L23.3776 22.3631V0L0.926758 12.9624V38.8855L23.3776 25.9231V48.2862L0.926758 61.2487L23.3776 74.2111L45.8284 61.2487V42.0429C45.8284 37.8855 43.6101 34.0447 40.0108 31.966Z"
-        fill="#F5F5F5"
-      />
-    </svg>
-  ),
-});
-
-export { LogoIcon };
