+++ src/utils/constants.ts
@@ -1,297 +1,3 @@
 export enum ENetworks {
   BETA_5 = 'https://beta-5.fuel.network/graphql',
 }
-
-export const termsOfUse = [
-  {
-    title: '1. The Bako Safe',
-    paragraphs: [
-      `The “Bako Safe” is a software application developed by Bako Labs LLC, a Brazilian company (“Bako” or “we”) for the purpose of managing private-public key pairs, blockchain addresses, and other data connected to user-controlled wallets. This agreement is between you and Bako.`,
-
-      `The Bako Safe interacts with a single blockchain network: the “Fuel Network,” an execution layer for modular blockchains.`,
-
-      `Bako Safe and the Fuel Network are in “testnet” form only, meaning they are live testing environments but are not the final version of the products. Testnets are designed to improve the operational experience, security, and design of the Fuel mainnet protocol and to expose potential issues and to improve participant experiences prior to the mainnet launch.`,
-    ],
-  },
-  {
-    title: '2. Licence To Use',
-    paragraphs: [
-      `Provided that you abide by the terms of this agreement, Bako grants you a personal, non-revocable, non-transferrable, non-sublicensable, worldwide licence to use Bako Safe.`,
-    ],
-  },
-  {
-    title: '3. Eligibility To Use',
-    paragraphs: [
-      `You acknowledge that you meet the following requirements to use Bako Safe:`,
-
-      `• Age of majority in your jurisdiction; and,`,
-
-      `• Intend to use Bako Safe for your own personal use and not for the benefit of a third party; and,`,
-
-      `• Understand the nature of the Fuel Network, blockchain wallets, and blockchain transactions; and,`,
-
-      `• Capable of securing your digital device(s) used with Bako Safe; and,`,
-
-      `• Not accessing Bako Safe from a territory subject to comprehensive economic sanctions by Brazil or other relevant jurisdictions.`,
-
-      `If at any time you are no longer eligible to use Bako Safe, according to the above criteria, then you must discontinue using Bako Safe.`,
-    ],
-  },
-  {
-    title: '4. Developer-Use Only: Non-Financial Transactions',
-    paragraphs: [
-      `The Fuel Network does not currently support sending or receiving valuable cryptocurrency or any form of blockchain token that can be used to transfer value to another person. You should only use Bako Safe if you understand that it is for testing the Fuel Network, and not for sending or receiving value.`,
-
-      `In your use of the testnet, you may accumulate “testnet tokens,” such as through a faucet, which are not, and shall never convert to or accrue to become mainnet tokens or any other tokens or virtual assets. Testnet tokens are virtual items with no monetary value. Testnet tokens do not constitute any currency or property of any type and are not redeemable, refundable, or eligible for any fiat or virtual currency or anything else of value. Testnet tokens are not transferable between users outside of the testnet, and you may not attempt to sell, trade, or transfer any testnet tokens outside of the testnet, or obtain any manner of credit using any testnet tokens. Any attempt to sell, trade, or transfer any testnet tokens outside of the testnet will be null and void.`,
-
-      `Please take care while using Bako Safe to ensure that you do not attempt to send valuable tokens (using third-party software) to an address generated by Bako Safe.`,
-    ],
-  },
-  {
-    title: '5. Browser-Based Access Terms',
-    paragraphs: [
-      `You acknowledge that Bako Safe is offered directly by Bako and is accessible through your browser without requiring installation or use of any third-party app stores (such as Apple or Google). Bako is solely responsible for providing support for Bako Safe and for any product liability, infringement of third-party intellectual property rights, failure to conform to laws or regulatory requirements, or any other issues related to Bako Safe.`,
-
-      `Since Bako Safe is not distributed through any app store, there are no app store terms or third-party agreements applicable to your use of Bako Safe. All interactions with the service are governed solely by this Agreement.`,
-    ],
-  },
-  {
-    title: '6. Your Security: Passkeys and External Owned Accounts (EOA)',
-    paragraphs: [
-      `Bako Safe does not use backup phrases. Instead, Bako Safe relies on passkeys stored on the users’ devices or External Owned Accounts (EOA) wallets, such as Fuel or Fuelet. These passkeys are managed by the user’s device and/or external wallets, and Bako cannot be responsible for their storage, recovery, or security.`,
-
-      `Since Bako Safe does not generate or manage backup phrases, it is crucial for users to ensure the security of their devices and associated passkeys. Bako Safe cannot recover passkeys or any associated access credentials. Therefore, users are solely responsible for safeguarding their passkeys and linked wallets, as any loss of access to these credentials may result in the inability to access your Bako Safe account or associated assets.`,
-    ],
-  },
-  {
-    title: '7. Technical Support',
-    paragraphs: [
-      `You may obtain technical information or advice from Bako if you are provided with technical support that you have requested. You are not required to make use of Bako's complimentary technical support, and Bako may discontinue offering such support to you (and anyone else) at any time. We do not charge a fee for technical support.
-      If Bako does provide you with technical support then you agree that any information provided shall be without warranty of any kind and you are solely responsible for any negative consequences that may occur as a result of misunderstanding, misapplying, or even accurately following the instructions or information provided by technical support. You agree to hold Bako harmless against any damages that occur, and you agree to indemnify Bako against any third-party claims made in relation to technical support that you obtain from Bako.
-      Technical support from Bako is not a substitute for retaining your own technical expert.
-      Bako will never instruct you to provide passkeys, or any other confidential information that should never leave your personal possession. We urge you to be cautious when obtaining technical support from third parties, or from people impersonating Bako through online scams. Please take care to obtain technical support through official channels only.`,
-    ],
-  },
-  {
-    title: '8. Fuel Network Disclaimer',
-    paragraphs: [
-      `The Bako Safe is independent of the Fuel Network. If you use the Fuel Network (or any other blockchain network) then you acknowledge that Bako may not be able to provide technical support or assist you with any issues that you encounter.`,
-    ],
-  },
-  {
-    title: '9. API Disclaimer',
-    paragraphs: [
-      `The Bako Safe depends on backend services (APIs) that may not always function correctly or be available. Although Bako takes care to ensure continuous operation of Bako Safe, you should be aware that downtime may occur and features may be updated which may cause technical issues. Bako is not responsible for any delays, errors, or problems that you may encounter due to service unavailability.`,
-    ],
-  },
-  {
-    title: '10. Third-Party Data and Pricing',
-    paragraphs: [
-      `Bako Safe incorporates third-party data, such as price information, that may not be reliable. You should always double-check any data that you receive through the Bako Safe using independent sources. Fuel is not responsible for any errors, whether due to display bugs or incorrect APIs (or any other cause) in relation to third-party data displayed within Bako Safe (even if not labelled as being “third-party data”).`,
-    ],
-  },
-  {
-    title: '11. No Warranty',
-    paragraphs: [
-      `To the maximum extent permitted by law, Bako Safe is provided on an “as-is”, “as-available” basis.`,
-
-      `Bako Safe is provided to you without any warranty of any kind, to the maximum extent permitted by law. The foregoing warranty disclaimer is applicable without regard to the type of warranty, whether fitness for a specific purpose, merchantability, or any other warranties imposed by law (whether common law, statute, or any other source of law).`,
-    ],
-  },
-  {
-    title: '12. Limitation Of Liability',
-    paragraphs: [
-      `Notwithstanding any term of this agreement, Bako’s liability to you shall be limited to the total amount paid by you, if any, for the use of Bako Safe. In no event shall Bako be liable for any indirect, incidental, special, punitive, or consequential damages, including but not limited to lost profits, loss of data, or business interruption, arising out of or in connection with your use or inability to use Bako Safe, even if Bako has been advised of the possibility of such damages.`,
-
-      `Furthermore, you acknowledge that Bako is not responsible for any loss of digital assets (whether valuable or otherwise) that you may experience due to the use or misuse of Bako Safe.`,
-
-      `This limitation of liability does not exclude or limit Bako’s liability in cases of willful misconduct (dolo) or gross negligence (culpa grave), in accordance with Brazilian law. You acknowledge that this limitation of liability is an essential part of this agreement, and that Bako would not be able to offer Bako Safe without your acceptance of these terms.`,
-    ],
-  },
-  {
-    title: '13. Indemnification',
-    paragraphs: [
-      `You shall indemnify Bako and its contractors, officers, directors, affiliates, and service providers against any claims brought by you (or by a third party in relation to your use of Bako Safe) for which the loss claimed for is partly or entirely caused by your misuse of Bako Safe or breach of this agreement.`,
-    ],
-  },
-  {
-    title: '14. Feedback',
-    paragraphs: [
-      `You may use the Bako Safe without providing any feedback to Bako. If you provide any feedback in relation to Bako Safe then you acknowledge that Bako may implement your idea, concept, or suggestion without any amounts being due to you.`,
-    ],
-  },
-  {
-    title: '15. Telemetry',
-    paragraphs: [
-      `Bako Safe makes use of industry-standard telemetry, which Bako uses to improve the product and ensure a quality experience for users. Bako telemetry does not include personally identifying information about you or other users. Bako telemetry is collected, processed, and used according to Bako Wallet Privacy Policy.`,
-    ],
-  },
-  {
-    title: '16. Amendment',
-    paragraphs: [
-      `Bako may amend this agreement with five business days’ notice. If you disagree with any amendments, your sole remedy is to cease using the Bako Safe.`,
-    ],
-  },
-  {
-    title: '17. Assignment',
-    paragraphs: [
-      `You may not assign this agreement. Bako may assign this agreement as part of a sale or restructuring.`,
-    ],
-  },
-  {
-    title: '18. Binding Agreement',
-    paragraphs: [
-      `This contract is a binding electronic agreement between the parties.`,
-    ],
-  },
-  {
-    title: '19. Legal Notice',
-    paragraphs: [
-      `You may provide notice to Bako by writing to: legal@bako.global.`,
-    ],
-  },
-  {
-    title: '20. Choice Of Forum',
-    paragraphs: [
-      `Any disputes related to this agreement shall be heard in the courts of [Placeholder for forum jurisdiction information—must specify courts within Brazil]. You may not bring a claim in any other jurisdiction and agree to promptly discontinue any actions pursued in any other jurisdiction.`,
-    ],
-  },
-  {
-    title: '21. Governing Law',
-    paragraphs: [
-      `This agreement, and any disputes related to it, shall be governed by the laws of [Placeholder for governing law information—should specify that Brazilian law governs this agreement]`,
-    ],
-  },
-  {
-    title: '22. Severability',
-    paragraphs: [
-      `Any provision or clause of this agreement, if void, invalid, unlawful, or unenforceable, shall be read down, or if it is not possible to do so while respecting the intention of the parties, severed from this agreement (to the minimum extent necessary).`,
-    ],
-  },
-  {
-    title: '23. Entire Agreement',
-    paragraphs: [
-      `This is the entire agreement between the parties and it supersedes any other understandings, agreements, or representations.`,
-
-      `Each party mutually acknowledges that they have entered into this agreement without relying on any document, representation, or information that is not expressly provided for in this agreement.`,
-    ],
-  },
-];
-
-export const privacyPolicy = [
-  {
-    title: '1. Who We Are',
-    paragraphs: [
-      `The right to privacy is a fundamental human right. We, Bako Labs LLC., make products that respect your rights in the way that they operate and empower you to ensure that your privacy is respected by others.`,
-    ],
-  },
-  {
-    title: '2. How Bako Safe Works',
-    paragraphs: [
-      `Bako Safe is a browser-based application that enables you, the user, to interact with the Fuel blockchain network without requiring you to disclose personal information (such as your name, email address, or other identifiers commonly required for software products). Instead of relying on traditional seed phrases, Bako Safe uses passkeys stored securely on your device or linked to an External Owned Account (EOA) wallet, such as Fuel or Fuelet.`,
-
-      `When you first use Bako Safe, a passkey is created and securely stored on your device, or you can link an existing EOA wallet. Unlike seed phrases, passkeys are not manually managed, and Bako Safe does not have access to or control over them. It is important to ensure the security of your device and EOA wallet, as Bako Safe cannot recover lost passkeys or restore access to your account.`,
-
-      `We encourage you to read this privacy policy in conjunction with other advice on using blockchain networks safely, securely, and privately.`,
-    ],
-  },
-  {
-    title: '3. What We Collect and Store',
-    paragraphs: [
-      `Bako Safe connects to backend server(s) that interact with the Fuel blockchain network. You may use a backend server that you provide, or Bako may provide one for you to use. If you use a backend server operated by Bako, this privacy policy applies to the use of that server as well (the “Bako Backend”).`,
-
-      `We collect certain incidental data as part of your use of the Bako Backend (if you use it), but we do not collect personal data such as IP addresses or other device identifiers. We do not perform analytics or use any third-party analytics providers. The data we collect is strictly related to providing the backend service that allows Bako Safe to fetch and transmit information to the Fuel blockchain network. This data is sent by your device as part of standard HTTPS requests between Bako Safe and the backend servers that connect to the Fuel blockchain network (“Backend API Calls”).`,
-
-      `In addition to incidental data collected through Backend API Calls, Bako Safe may also store and transmit certain telemetry data, crash/bug data, and other logs (“Telemetry Data”) to help ensure the quality and functionality of Bako Safe. Telemetry Data does not include personal information and does not include your passkey(s), public key(s), or any similar sensitive data.`,
-    ],
-  },
-  {
-    title: '4. Maximizing Your Privacy',
-    paragraphs: [
-      `Blockchain networks are designed to operate in a decentralized manner, and Bako Safe can be used to enhance this aspect if you so choose. For better privacy, you can use your own backend server that you control (instead of the Bako Backend, as mentioned above). We encourage you to run your own backend server for your transactions to better protect your privacy, ensure continuity of service, and take full advantage of decentralized software.`,
-
-      `Although we do not store your name, address, email address, phone number, or other personal identifiers that would link your identity to your use of Bako Safe, it is possible that your transaction patterns on the Fuel blockchain network, such as using addresses known by others to be associated with you, could reveal information about your usage.`,
-
-      `We encourage users to be mindful of their blockchain activity and to use modern precautions to help protect their privacy.`,
-    ],
-  },
-  {
-    title: '5. Optional User Support Service',
-    paragraphs: [
-      `We may offer technical support to users like you. However, using our support services is not required to use Bako Safe.`,
-
-      `If we provide support and you choose to use it, please be aware that this may involve sharing your email address or other contact information that could be associated with your use of Bako Safe. Even if we do not offer direct technical support, you may still contact us through other channels, and this privacy policy will apply to any such communications.`,
-
-      `Our support systems, to the extent offered, may retain records of interactions to provide a written record of issues. These records may be kept for liability purposes, training, or to ensure continuity of support.`,
-    ],
-  },
-  {
-    title: '6. Collection Purposes',
-    paragraphs: [
-      `In addition to what is described elsewhere in this policy, we collect personal and other information for the following purposes:`,
-      `• Improving Bako Safe; or,`,
-      `• Providing Bako Safe and Backend API Calls to you; or,`,
-      `• Understanding how users interact with Bako Safe to improve the service; or,`,
-      `• Providing technical support services; or,`,
-      `• Ensuring that we only respond to data access requests from the person to whom the data request relates; or,`,
-      `• Debugging and resolving technical issues with the Bako Safe product.`,
-    ],
-  },
-  {
-    title: '7. Personal Data And Sensitive Personal Data',
-    paragraphs: [
-      `We do not collect any “personal data” or “sensitive personal data”, as defined in applicable legislation in Brazil. We strongly discourage users from sending sensitive personal data to us by way of email, user support ticket, mail, or other means.`,
-    ],
-  },
-  {
-    title: '8. How We Handle Personal Data',
-    paragraphs: [
-      `We may, in some circumstances, collect personal data that relates to you or could be related to you in the future. Personal data, including data that could become personal data if correlated with other information held by us or others, is protected by security measures such as encryption, permissions, and “need to know” access. All personnel (whether employees or contractors) who have access to personal data are contractually obligated to protect it.`,
-    ],
-  },
-  {
-    title: '9. Blockchain Networks',
-    paragraphs: [
-      `Bako Safe enables you to create blockchain transactions on the Fuel blockchain network. Blockchain transactions are generally permanent and immutable. We do not control, and cannot alter, the Fuel blockchain network. You should take care to only create transactions that you wish to be public and permanent.`,
-    ],
-  },
-  {
-    title: '10. Data Retention',
-    paragraphs: [
-      `Data that we collected is retained until the data is no longer useful for the purpose(s) for which the data was collected. Some data may be of a transient nature, such as automatically generated logs for cloud services, and may be deleted shortly after being generated, whereas other data may be retained for years. In each case, we consider the type of data, collected purpose, and current need.`,
-    ],
-  },
-  {
-    title: '11. Our Website(s)',
-    paragraphs: [
-      `You may obtain information related to Bako Safe from websites operated by Bako or other legal entities (such as affiliates), including our social media channels (together, “Bako Websites”). The Bako Websites may use cookies or other persistent identifiers and may employ analytics software to understand user preferences and improve our services.`,
-
-      `It is important to review the privacy policies applicable to these Bako Websites to understand their impact on your privacy. Since Bako Safe is a browser-based application, there is no need to download software from an app store, minimizing the need for third-party involvement in accessing the service.`,
-    ],
-  },
-  {
-    title: '12. Government Requests',
-    paragraphs: [
-      `We abide by the law, while balancing our legal obligation with the right to privacy that users have. We may provide information to law enforcement or other government authorities in Brazil (or elsewhere) upon receiving a production order or other lawful request for data.`,
-    ],
-  },
-  {
-    title: '13. Data Access Request',
-    paragraphs: [
-      `You may write to us in order to obtain access to the personal data that we collect and store that relates to you: legal@bako.global.`,
-    ],
-  },
-  {
-    title: '14. Right To Correct Records',
-    paragraphs: [
-      `If you believe that information (of any kind) held by us may be inaccurate or incomplete then you may write to us in order to correct the data: legal@bako.global.`,
-    ],
-  },
-  {
-    title: '15. Information Commissioner And Your Rights',
-    paragraphs: [
-      `We are based in Brazil, and the laws of Brazil apply to our operations.
-You may write to Autoridade Nacional de Proteção de Dados (ANPD) if you have a complaint regarding our handling or processing of data. We encourage you to write to us first, in order to resolve any issues: legal@bako.global.`,
-    ],
-  },
-];
